Improve node ble (#409)
* Reset the isStarted flag during the reset procedure

* Always return a success response when requesting control, even if control is already given.

This seems to resolve an issue i have with FortiusANT and Zwift running on AppleTV or iPhone. If ControlNotPermitted was returned, Zwift would continuously spam with RequestControl commands trying to take control, even if it already had control. This prevented the ERG mode to work correctly because Zwift will not send the target power until it has control. Probably this has been an issue for a long time, since Zwift updated their bluetooth stack beginning of january 2022.